solve_network refactor, ERM & opts testing  (#592)
* fix filter components for single horizon

* store erm duals

* write tests and refactor

* fix tests, still errors on reserves

* fix filter components for single horizon

* store erm duals

* write tests and refactor

* fix tests, still errors on reserves

* Functionalize brownfield for myopic

* Fix ERM and test for ERM

* remove defunct AI writen tests

* ERM link support, all timesteps, rm duplicate function

* Address comments

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* docs to run test

---------
