+++ workflow/scripts/test/fixtures/technology_capacity_targets.csv
@@ -1,0 +1,4 @@
+name,region,carrier,planning_horizon,min,max
+wind_west_min,west,onwind,2030,300,
+wind_west_max,west,onwind,2030,,500
+battery_west_min,west,battery,2030,100,
